move value filter parse date to utils, config fixes for rds/rdscluster
